can draw a bit of chart now

+++ BookstoreApi/Management/Models/RevenueByDay.cs
@@ -5,6 +5,6 @@ namespace Mangement.Models
     public class RevenueByDay
     {
         public DateTime Date { get; set; }
-        public decimal TotalRevenue { get; set; }
+        public double TotalRevenue { get; set; }
     }
 }
